+++ classes/mode-interface.php
@@ -9,8 +9,6 @@ namespace ARI;
 interface Mode_Interface {
 	public function render_image( $html = '' );
 
-	public function add_filters();
-
 	public function set_args( $args );
 
 	public function set_attachment_id( $id );